refactor: updating the auth_serv name to AuthenticationService
